feat: Disable individual accounts from syncing
And be able to delete all transactions

I added my a new bank connection in SimpleFin that has joint banking
accounts with an existing bank connection. This lead to duplicate
transactions showing up from SimpleFin. A relatively quick way to deal
with this is to disable syncing of particular accounts and remove any
transactions that were part of that account.

The account still comes through in SimpleFin, so don't save the updated
balance or any transactions when fetched.

A future net-worth display should only look for balances within some
recent time period / only active accounts