fix: Reduce duplicate rows in amount queries
Which should give a more correct value

Transactions with multiple labels that match (income and income.paycheck as an example when searching for income) would be included once for each label. Don't do that!

These queries could still use deduplication to match the rest of the
query filter logic to use a single query based on the result needed, not
the conditions