fix: Generate querystrings that can be parsed
The axum-extra query extractor was not able to parse the to_querystring
result appropriately. The extractor looked for lists like a=1 a=2 a=3 to
mean list a, with 3 elements.

serde_qs would generate that as a[0]=1 a[1]=2 a[2]=3.

Instead, use the same library internally that axum-extra does. This
should ideally test with a request... but I added at least some tests!